+++ filament/backend/src/opengl/PlatformEGL.h
@@ -65,6 +65,9 @@ public:
             uint32_t w, uint32_t h, backend::TextureFormat format) noexcept final;
     void destroyExternalTextureStorage(ExternalTexture* ets) noexcept final;
 
+    void createExternalImageTexture(void* texture) noexcept final;
+    void destroyExternalImage(void* texture) noexcept final;
+
     int getOSVersion() const noexcept final;
 
 private:
@@ -78,6 +81,11 @@ private:
     EGLConfig mEGLConfig;
     EGLConfig mEGLTransparentConfig;
     int mOSVersion;
+
+    // supported extensions detected at runtime
+    struct {
+        bool OES_EGL_image_external_essl3 = false;
+    } ext;
 
     ExternalStreamManagerAndroid& mExternalStreamManager;
     ExternalTextureManagerAndroid& mExternalTextureManager;